requirements.txt
anthropic-agent-sdk>=0.1.0
ccxt>=4.2.0
pandas>=2.0.0
pandas-ta>=0.3.14b
python-dotenv>=1.0.0
aiosqlite>=0.19.0
click>=8.1.0
rich>=13.0.0
